# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UAV Platform API
  version: 2.0.0
  description: |
    基于WRF气象驱动的无人机VRP智能路径规划系统 API 平台。

    ## 功能模块
    - **平台管理**: 租户管理、API密钥管理、用量统计
    - **气象数据**: WRF/风云/天姿/风雷多源气象数据查询与融合
    - **数据同化**: 3DVAR/4DVAR/5DVAR/EnKF/Hybrid/Enhanced Bayesian 同化算法
    - **风险评估**: 综合风险感知与区域风险栅格地图
    - **适航评估**: 全维度适航评估与适航标准查询
    - **观测决策**: 自适应/计划/应急观测任务管理与决策
    - **路径规划**: VRPTW/DE-RRT*/DWA/MPC/A*/Dijkstra/RRT* 路径规划
    - **UTM**: 空域管理、飞行计划、实时追踪与冲突检测

    ## 认证方式
    - **Bearer Token**: JWT 认证，用于平台管理接口
    - **API Key**: HMAC 签名认证，用于业务接口调用

  contact:
    name: UAV Platform Team
    email: support@uav-platform.example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: 开发环境（通过网关访问）
  - url: https://api.uav-platform.example.com
    description: 生产环境
...
paths:
...
  /api/v1/platform/api-keys:
    post:
      tags: [Platform]
      summary: 生成API密钥
      description: 为指定租户生成新的API密钥，支持配置速率限制和有效期。
      operationId: generateApiKey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateApiKeyRequest'
            example:
              tenantId: 1
              name: "生产环境密钥"
              rateLimit: 1000
              expiresInDays: 365
      responses:
        '200':
          description: 密钥生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultApiKey'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/platform/api-keys/{id}:
    get:
      tags: [Platform]
      summary: 获取API密钥详情
      description: 根据ID获取API密钥详细信息。
      operationId: getApiKey
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultApiKey'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Platform]
      summary: 删除API密钥
      description: 删除指定的API密钥。
      operationId: deleteApiKey
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/platform/api-keys/{id}/enable:
    post:
      tags: [Platform]
      summary: 启用API密钥
      description: 启用已禁用的API密钥。
      operationId: enableApiKey
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
  /api/v1/platform/api-keys/{id}/disable:
    post:
      tags: [Platform]
      summary: 禁用API密钥
      description: 禁用指定的API密钥。
      operationId: disableApiKey
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
  /api/v1/platform/api-keys/tenant/{tenantId}:
    get:
      tags: [Platform]
      summary: 查询租户的API密钥列表
      description: 获取指定租户下所有API密钥。
      operationId: listApiKeysByTenant
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListApiKey'
...
  /api/v1/observation/tasks:
    get:
      tags: [Observation]
      summary: 列出所有观测任务
      description: 获取当前租户下所有观测任务列表。
      operationId: listObservationTasks
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListObservationTask'

    post:
      tags: [Observation]
      summary: 创建观测任务
      description: 创建新的观测任务，支持自适应(ADAPTIVE)、计划(PLANNED)和应急(EMERGENCY)三种类型。
      operationId: createObservationTask
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObservationRequest'
            example:
              type: "ADAPTIVE"
              targetArea: '{"minLon":116.0,"minLat":39.8,"maxLon":116.5,"maxLat":40.0}'
              sensorType: "LIDAR"
              priority: 5
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultObservationTask'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/observation/tasks/{id}:
    get:
      tags: [Observation]
      summary: 获取观测任务详情
      description: 根据ID获取观测任务详细信息。
      operationId: getObservationTask
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultObservationTask'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/observation/tasks/{id}/status:
    post:
      tags: [Observation]
      summary: 更新观测任务状态
      description: 更新指定观测任务的执行状态。
      operationId: updateObservationTaskStatus
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: status
          in: query
          required: true
          description: 目标状态
          schema:
            type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultObservationTask'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # 观测决策
  # ============================================================
...
  /api/v1/planning/path:
    post:
      tags: [Planning]
      summary: 提交路径规划任务
      description: 提交单条路径规划任务，支持多种优化目标（风险最小、能耗最低、时间最短、综合平衡）。
      operationId: submitPathPlanning
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanPathRequest'
            example:
              start: {"lon": 116.3, "lat": 39.9, "alt": 100}
              end: {"lon": 116.5, "lat": 40.0, "alt": 150}
              waypoints:
                - {"lon": 116.4, "lat": 39.95, "alt": 120}
              uavModel: "DJI-M300"
              optimizationTarget: "BALANCED"
              constraints:
                maxAltitude: 200
                minAltitude: 50
      responses:
        '200':
          description: 任务提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPlanningTask'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/planning/mission:
    post:
      tags: [Planning]
      summary: 提交任务规划
      description: 提交多无人机多任务规划，使用VRPTW算法进行任务分配和路径优化。
      operationId: submitMissionPlanning
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanMissionRequest'
            example:
              uavList:
                - {"id": "UAV-001", "model": "DJI-M300", "battery": 5200}
              taskList:
                - {"id": "T-001", "location": {"lon": 116.4, "lat": 39.9}, "priority": 1}
              areaBounds: {"minLon": 116.0, "minLat": 39.8, "maxLon": 116.5, "maxLat": 40.0}
      responses:
        '200':
          description: 任务提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPlanningTask'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/planning/tasks:
    get:
      tags: [Planning]
      summary: 列出所有规划任务
      description: 获取当前租户下所有路径规划和任务规划列表。
      operationId: listPlanningTasks
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListPlanningTask'
  /api/v1/planning/tasks/{id}:
    get:
      tags: [Planning]
      summary: 获取规划任务状态
      description: 根据ID查询规划任务的当前状态和进度。
      operationId: getPlanningTask
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPlanningTask'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/planning/tasks/{id}/result:
    get:
      tags: [Planning]
      summary: 获取路径规划结果
      description: 获取已完成路径规划任务的航点列表、总距离、预计耗时、风险评分和能耗估算。
      operationId: getPathResult
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPathResult'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/planning/tasks/{id}/mission:
    get:
      tags: [Planning]
      summary: 获取任务规划结果
      description: 获取已完成任务规划的调度方案和整体评分。
      operationId: getMissionPlan
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultMissionPlan'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/planning/tasks/{id}/cancel:
    post:
      tags: [Planning]
      summary: 取消规划任务
      description: 取消正在运行或排队中的规划任务。
      operationId: cancelPlanningTask
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # UTM - 空域管理
  # ============================================================
...
components:
...
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      description: 任务ID
      schema:
        type: integer
        format: int64
...
    ApiKeyId:
      name: id
      in: path
      required: true
      description: API密钥ID
      schema:
        type: integer
        format: int64

  # ----------------------------------------------------------
  # 数据模型
  # ----------------------------------------------------------
  schemas:
...
    Result:
      type: object
      description: 统一API响应结构
      properties:
        code:
          type: integer
          description: 业务状态码 (200=成功)
          example: 200
        message:
          type: string
          description: 提示消息
          example: success
        data:
          description: 响应数据（泛型）
        requestId:
          type: string
          description: 请求追踪ID
        timestamp:
          type: integer
          format: int64
          description: 时间戳
    ResultVoid:
      allOf:
        - $ref: '#/components/schemas/Result'
      description: 无数据的成功响应
...
    GenerateApiKeyRequest:
      type: object
      required: [tenantId, name]
      properties:
        tenantId:
          type: integer
          format: int64
        name:
          type: string
          description: 密钥名称
        rateLimit:
          type: integer
          description: 速率限制(次/分钟)
        expiresInDays:
          type: integer
          description: 有效天数
    ApiKey:
      type: object
      description: API密钥实体
      properties:
        id:
          type: integer
          format: int64
        tenantId:
          type: integer
          format: int64
        keyValue:
          type: string
          description: API Key值
        secret:
          type: string
          description: 密钥Secret
        name:
          type: string
        status:
          type: integer
          description: "状态: 1=启用, 0=禁用"
        rateLimit:
          type: integer
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    ResultApiKey:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/ApiKey'
    ResultListApiKey:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApiKey'

    # --- 气象数据 ---
...
    CreateObservationRequest:
      type: object
      required: [type, targetArea, sensorType, priority]
      properties:
        type:
          type: string
          description: "任务类型: ADAPTIVE, PLANNED, EMERGENCY"
        targetArea:
          type: string
          description: 目标区域JSON
        sensorType:
          type: string
          description: 传感器类型
        priority:
          type: integer
          description: 优先级(1-10)
    ObservationTask:
      type: object
      description: 观测任务
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          description: "任务类型: ADAPTIVE, PLANNED, EMERGENCY"
        status:
          type: string
          description: 任务状态
        sensorConfigJson:
          type: string
          description: 传感器配置JSON
        plannedPathJson:
          type: string
          description: 规划路径JSON
        actualPathJson:
          type: string
          description: 实际路径JSON
        dataQuality:
          type: number
          format: double
          description: 数据质量评分(0-100)
        assimilationFeedbackJson:
          type: string
          description: 同化反馈JSON
        tenantId:
          type: string
        createdAt:
          type: string
          format: date-time
    ResultObservationTask:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/ObservationTask'
    ResultListObservationTask:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ObservationTask'
...
    PlanPathRequest:
      type: object
      required: [start, end, optimizationTarget]
      properties:
        start:
          type: object
          description: 起点坐标 {lon, lat, alt}
          required: [lon, lat, alt]
          properties:
            lon:
              type: number
              format: double
            lat:
              type: number
              format: double
            alt:
              type: number
              format: double
        end:
          type: object
          description: 终点坐标 {lon, lat, alt}
          required: [lon, lat, alt]
          properties:
            lon:
              type: number
              format: double
            lat:
              type: number
              format: double
            alt:
              type: number
              format: double
        waypoints:
          type: array
          description: 途经点列表
          items:
            type: object
            properties:
              lon:
                type: number
                format: double
              lat:
                type: number
                format: double
              alt:
                type: number
                format: double
        uavModel:
          type: string
          description: 无人机型号
        constraints:
          type: object
          additionalProperties:
            type: object
          description: 约束条件
        optimizationTarget:
          type: string
          description: "优化目标: RISK, ENERGY, TIME, BALANCED"
    PlanMissionRequest:
      type: object
      required: [uavList, taskList]
      properties:
        uavList:
          type: array
          description: 无人机列表
          items:
            type: object
            additionalProperties:
              type: object
        taskList:
          type: array
          description: 任务列表
          items:
            type: object
            additionalProperties:
              type: object
        areaBounds:
          type: object
          description: 区域边界 {minLon, minLat, maxLon, maxLat}
          properties:
            minLon:
              type: number
              format: double
            minLat:
              type: number
              format: double
            maxLon:
              type: number
              format: double
            maxLat:
              type: number
              format: double
        priorities:
          type: object
          additionalProperties:
            type: integer
          description: 优先级配置
    PlanningTask:
      type: object
      description: 规划任务
      properties:
        id:
          type: integer
          format: int64
        algorithmType:
          type: string
          description: "算法类型: VRPTW, DERRTSTAR, DWA, MPC, A_STAR, DIJKSTRA, RRTSTAR"
        status:
          type: string
          description: "任务状态: QUEUED, RUNNING, SUCCESS, FAILED, TIMEOUT, CANCELLED"
        paramsJson:
          type: string
          description: 算法参数JSON
        resultJson:
          type: string
          description: 结果数据JSON
        progress:
          type: integer
          description: 进度(0-100)
        errorMsg:
          type: string
          description: 错误信息
        tenantId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    ResultPlanningTask:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/PlanningTask'
    ResultListPlanningTask:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PlanningTask'
    PathResult:
      type: object
      description: 路径规划结果
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        waypointsJson:
          type: string
          description: 航点列表JSON [{lon, lat, alt, speed, time}]
        totalDistance:
          type: number
          format: double
          description: 总距离(m)
        estimatedTime:
          type: number
          format: double
          description: 预计耗时(s)
        riskScore:
          type: number
          format: double
          description: 风险评分(0-100)
        energyConsumption:
          type: number
          format: double
          description: 能耗估算(Wh)
    ResultPathResult:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/PathResult'
    MissionPlan:
      type: object
      description: 任务规划结果
      properties:
        id:
          type: integer
          format: int64
        uavsJson:
          type: string
          description: 无人机列表JSON
        tasksJson:
          type: string
          description: 任务列表JSON
        scheduleJson:
          type: string
          description: 调度方案JSON
        overallScore:
          type: number
          format: double
          description: 整体评分
        tenantId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    ResultMissionPlan:
      allOf:
        - $ref: '#/components/schemas/Result'
      properties:
        data:
          $ref: '#/components/schemas/MissionPlan'

    # --- UTM ---
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证，用于平台管理接口

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key 认证，配合 HMAC 签名使用

  # ----------------------------------------------------------
  # 通用响应
  # ----------------------------------------------------------
  responses:
    Success:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResultVoid'
          example:
            code: 200
            message: success
            data: null
            timestamp: 1717200000000
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1000
            message: "请求参数错误"
            data: null
            timestamp: 1717200000000
    Unauthorized:
      description: 未授权，请先登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1001
            message: "未授权，请先登录"
            data: null
            timestamp: 1717200000000
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1002
            message: "权限不足"
            data: null
            timestamp: 1717200000000
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Result'
          example:
            code: 1003
            message: "资源不存在"
            data: null
            timestamp: 1717200000000
...